+++ tests/test_github_info_loader.py
@@ -1,4 +1,4 @@
-from leafhopper.descriptors.github_info_loader import (
+from leafhopper.descriptors.extra.github_info_loader import (
     GithubInfoLoader,
     has_github_homepage,
 )
@@ -7,8 +7,8 @@ GITHUB_REPO_HOMEPAGE = "https://github.com/twbs/bootstrap"
 
 
 def test_github_info_loader():
-    loader = GithubInfoLoader(GITHUB_REPO_HOMEPAGE)
-    repo = loader.load()
+    loader = GithubInfoLoader()
+    repo = loader.load({"name": "bootstrap", "homepage": GITHUB_REPO_HOMEPAGE})
     assert repo["name"] == "bootstrap"
     assert "license" in repo
     assert "description" in repo