+++ CHANGELOG.md
@@ -1,5 +1,9 @@
+# 0.3.0 - 2022-07-17
+* support maven project type
+# 0.2.1 - 2022-07-16
+* improve documentation
 # 0.2.0 - 2022-07-16
-* support for poetry project type
+* support poetry project type
 # 0.1.1 - 2022-07-16
 * support setting logging level via command line argument
 # 0.1.0 - 2022-07-16